requirements.txt
transformers==4.31.0
datasets==2.14.4
pytorch
numpy
